+++ libgroup.cpp
@@ -206,11 +206,11 @@ int32 scriptlib::group_filter_select(lua_State *L) {
 			cset.erase(pcard);
 	}
 	duel* pduel = pgroup->pduel;
-	uint32 playerid = lua_tointeger(L, 2);
+	uint32 playerid = (uint32)lua_tointeger(L, 2);
 	if(playerid != 0 && playerid != 1)
 		return 0;
-	uint32 min = lua_tointeger(L, 4);
-	uint32 max = lua_tointeger(L, 5);
+	uint32 min = (uint32)lua_tointeger(L, 4);
+	uint32 max = (uint32)lua_tointeger(L, 5);
 	uint32 extraargs = lua_gettop(L) - 6;
 	pduel->game_field->core.select_cards.clear();
 	for (auto& pcard : cset) {
@@ -244,11 +244,11 @@ int32 scriptlib::group_select(lua_State *L) {
 			cset.erase(pcard);
 	}
 	duel* pduel = pgroup->pduel;
-	uint32 playerid = lua_tointeger(L, 2);
+	uint32 playerid = (uint32)lua_tointeger(L, 2);
 	if(playerid != 0 && playerid != 1)
 		return 0;
-	uint32 min = lua_tointeger(L, 3);
-	uint32 max = lua_tointeger(L, 4);
+	uint32 min = (uint32)lua_tointeger(L, 3);
+	uint32 max = (uint32)lua_tointeger(L, 4);
 	pduel->game_field->core.select_cards.clear();
 	for (auto& pcard : cset) {
 		pduel->game_field->core.select_cards.push_back(pcard);
@@ -273,7 +273,7 @@ int32 scriptlib::group_select_unselect(lua_State *L) {
 	group* pgroup1 = *(group**)lua_touserdata(L, 1);
 	group* pgroup2 = *(group**)lua_touserdata(L, 2);
 	duel* pduel = pgroup1->pduel;
-	uint32 playerid = lua_tointeger(L, 3);
+	uint32 playerid = (uint32)lua_tointeger(L, 3);
 	if(playerid != 0 && playerid != 1)
 		return 0;
 	if(pgroup1->container.size() + pgroup2->container.size() == 0)
@@ -296,11 +296,11 @@ int32 scriptlib::group_select_unselect(lua_State *L) {
 	}
 	uint32 min = 1;
 	if(lua_gettop(L) > 5) {
-		min = lua_tointeger(L, 6);
+		min = (uint32)lua_tointeger(L, 6);
 	}
 	uint32 max = 1;
 	if(lua_gettop(L) > 6) {
-		max = lua_tointeger(L, 7);
+		max = (uint32)lua_tointeger(L, 7);
 	}
 	if(min > max)
 		min = max;
@@ -332,8 +332,8 @@ int32 scriptlib::group_random_select(lua_State *L) {
 	check_param_count(L, 3);
 	check_param(L, PARAM_TYPE_GROUP, 1);
 	group* pgroup = *(group**) lua_touserdata(L, 1);
-	int32 playerid = lua_tointeger(L, 2);
-	uint32 count = lua_tointeger(L, 3);
+	int32 playerid = (int32)lua_tointeger(L, 2);
+	uint32 count = (uint32)lua_tointeger(L, 3);
 	duel* pduel = pgroup->pduel;
 	group* newgroup = pduel->new_group();
 	if(count > pgroup->container.size())
@@ -376,7 +376,7 @@ int32 scriptlib::group_is_exists(lua_State *L) {
 			cset.erase(pcard);
 	}
 	duel* pduel = pgroup->pduel;
-	uint32 count = lua_tointeger(L, 3);
+	uint32 count = (uint32)lua_tointeger(L, 3);
 	uint32 extraargs = lua_gettop(L) - 4;
 	uint32 fcount = 0;
 	uint32 result = FALSE;
@@ -398,9 +398,9 @@ int32 scriptlib::group_check_with_sum_equal(lua_State *L) {
 	check_param(L, PARAM_TYPE_FUNCTION, 2);
 	group* pgroup = *(group**) lua_touserdata(L, 1);
 	duel* pduel = pgroup->pduel;
-	int32 acc = lua_tointeger(L, 3);
-	int32 min = lua_tointeger(L, 4);
-	int32 max = lua_tointeger(L, 5);
+	int32 acc = (int32)lua_tointeger(L, 3);
+	int32 min = (int32)lua_tointeger(L, 4);
+	int32 max = (int32)lua_tointeger(L, 5);
 	if(min < 0)
 		min = 0;
 	if(max < min)
@@ -426,12 +426,12 @@ int32 scriptlib::group_select_with_sum_equal(lua_State *L) {
 	check_param(L, PARAM_TYPE_FUNCTION, 3);
 	group* pgroup = *(group**) lua_touserdata(L, 1);
 	duel* pduel = pgroup->pduel;
-	int32 playerid = lua_tointeger(L, 2);
+	int32 playerid = (int32)lua_tointeger(L, 2);
 	if(playerid != 0 && playerid != 1)
 		return 0;
-	int32 acc = lua_tointeger(L, 4);
-	int32 min = lua_tointeger(L, 5);
-	int32 max = lua_tointeger(L, 6);
+	int32 acc = (int32)lua_tointeger(L, 4);
+	int32 min = (int32)lua_tointeger(L, 5);
+	int32 max = (int32)lua_tointeger(L, 6);
 	if(min < 0)
 		min = 0;
 	if(max < min)
@@ -473,7 +473,7 @@ int32 scriptlib::group_check_with_sum_greater(lua_State *L) {
 	check_param(L, PARAM_TYPE_FUNCTION, 2);
 	group* pgroup = *(group**) lua_touserdata(L, 1);
 	duel* pduel = pgroup->pduel;
-	int32 acc = lua_tointeger(L, 3);
+	int32 acc = (int32)lua_tointeger(L, 3);
 	int32 extraargs = lua_gettop(L) - 3;
 	field::card_vector cv(pduel->game_field->core.must_select_cards);
 	int32 mcount = (int32)cv.size();
@@ -495,10 +495,10 @@ int32 scriptlib::group_select_with_sum_greater(lua_State *L) {
 	check_param(L, PARAM_TYPE_FUNCTION, 3);
 	group* pgroup = *(group**) lua_touserdata(L, 1);
 	duel* pduel = pgroup->pduel;
-	int32 playerid = lua_tointeger(L, 2);
+	int32 playerid = (int32)lua_tointeger(L, 2);
 	if(playerid != 0 && playerid != 1)
 		return 0;
-	int32 acc = lua_tointeger(L, 4);
+	int32 acc = (int32)lua_tointeger(L, 4);
 	int32 extraargs = lua_gettop(L) - 4;
 	pduel->game_field->core.select_cards.assign(pgroup->container.begin(), pgroup->container.end());
 	for(auto& pcard : pduel->game_field->core.must_select_cards) {
